+++ com.unity.render-pipelines.high-definition/Editor/Material/Hair/ShaderGraph/HairSettingsView.cs
@@ -101,12 +101,12 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline.Drawing
                     });
                 });
 
-                ps.Add(new PropertyRow(CreateLabel("Transparent Writes Velocity", indentLevel)), (row) =>
+                ps.Add(new PropertyRow(CreateLabel("Transparent Writes Motion Vector", indentLevel)), (row) =>
                 {
                     row.Add(new Toggle(), (toggle) =>
                     {
-                        toggle.value = m_Node.transparentWritesVelocity.isOn;
-                        toggle.OnToggleChanged(ChangeTransparentWritesVelocity);
+                        toggle.value = m_Node.transparentWritesMotionVec.isOn;
+                        toggle.OnToggleChanged(ChangeTransparentWritesMotionVec);
                     });
                 });
 
@@ -328,12 +328,12 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline.Drawing
             m_Node.alphaTestDepthPostpass = td;
         }
 
-        void ChangeTransparentWritesVelocity(ChangeEvent<bool> evt)
+        void ChangeTransparentWritesMotionVec(ChangeEvent<bool> evt)
         {
-            m_Node.owner.owner.RegisterCompleteObjectUndo("Transparent Writes Velocity Change");
-            ToggleData td = m_Node.transparentWritesVelocity;
+            m_Node.owner.owner.RegisterCompleteObjectUndo("Transparent Writes Motion Vector Change");
+            ToggleData td = m_Node.transparentWritesMotionVec;
             td.isOn = evt.newValue;
-            m_Node.transparentWritesVelocity = td;
+            m_Node.transparentWritesMotionVec = td;
         }
 
         void ChangeAlphaTestShadow(ChangeEvent<bool> evt)